requirement.txt
starknet-py
black
python-dotenv